+++ src/com/jdrago/blackout/bridge/NativeApp.java
@@ -1,6 +1,6 @@
 package com.jdrago.blackout.bridge;
 
 public interface NativeApp {
-    public void blit(String textureName, int srcX, int srcY, int srcW, int srcH, int dstX, int dstY, int dstW, int dstH, float rot, float anchorX, float anchorY);
+    public void blit(String textureName, float srcX, float srcY, float srcW, float srcH, float dstX, float dstY, float dstW, float dstH, float rot, float anchorX, float anchorY);
     public void log(String s);
 }
